requirements.txt

alpaca-trade-api              1.5.0
python-telegram-bot           13.11
pandas                        1.3.4
requests                      2.26.0
